Conflicts:
	app/models/ontology_version.rb

+++ spec/lib/hets_spec.rb
@@ -1,33 +1,32 @@
-require 'test_helper'
+require 'spec_helper'
 
-class HetsTest < ActiveSupport::TestCase
+describe Hets do
+
+  after do
+    File.delete @xml_path if @xml_path
+  end
+
   context 'Output directory parameter' do
-    setup do
+    before do
       @xml_path = Hets.parse Rails.root.join('test/fixtures/ontologies/owl/pizza.owl'), [], '/tmp'
     end
 
-    should 'correctly be used' do
+    it 'correctly be used' do
       assert @xml_path.starts_with? '/tmp'
-    end
-
-    teardown do
-      File.delete @xml_path
     end
   end
 
   %w(owl/pizza.owl owl/generations.owl clif/cat.clif).each do |path|
     context path do
-      setup do
-        assert_nothing_raised do
-          @xml_path = Hets.parse Rails.root.join("test/fixtures/ontologies/#{path}"), [], '/tmp'
-        end
+      before do
+        @xml_path = Hets.parse Rails.root.join("test/fixtures/ontologies/#{path}"), [], '/tmp'
       end
 
-      should 'have created output file' do
+      it 'have created output file' do
         assert File.exists? @xml_path
       end
 
-      should 'have generated importable output' do
+      it 'have generated importable output' do
         assert_nothing_raised do
           ontology = FactoryGirl.create :ontology
           user = FactoryGirl.create :user
@@ -35,28 +34,22 @@ class HetsTest < ActiveSupport::TestCase
           `git checkout #{@xml_path} 2>/dev/null`
         end
       end
-
-      teardown do
-        File.delete @xml_path
-      end
     end
   end
 
   context 'with url-catalog' do
-    setup do
-      assert_nothing_raised do
-        @xml_path = Hets.parse \
-          Rails.root.join("test/fixtures/ontologies/clif/monoid.clif"),
-          ["http://colore.oor.net/magma=file://#{Rails.root.join('test/fixtures/ontologies/clif')}"],
-          '/tmp'
-      end
+    before do
+      @xml_path = Hets.parse \
+        Rails.root.join("test/fixtures/ontologies/clif/monoid.clif"),
+        ["http://colore.oor.net=http://develop.ontohub.org/colore/ontologies"],
+        '/tmp'
     end
 
-    should 'have created output file' do
+    it 'have created output file' do
       assert File.exists? @xml_path
     end
 
-    should 'have generated importable output' do
+    it 'have generated importable output' do
       assert_nothing_raised do
         ontology = FactoryGirl.create :ontology
         user = FactoryGirl.create :user
@@ -64,13 +57,9 @@ class HetsTest < ActiveSupport::TestCase
         `git checkout #{@xml_path} 2>/dev/null`
       end
     end
-
-    teardown do
-      File.delete @xml_path
-    end
   end
 
-  should 'raise exception if provided with wrong file-format' do
+  it 'raise exception if provided with wrong file-format' do
     assert_raise Hets::HetsError do
       Hets.parse Rails.root.join('test/fixtures/ontologies/xml/valid.xml')
     end
